+++ src/main/java/com/kiranaservices/kirana_transactions/model/Report.java
@@ -3,6 +3,7 @@ package com.kiranaservices.kirana_transactions.model;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.util.Date;
 import java.util.List;
 
 @Getter
@@ -10,14 +11,14 @@ import java.util.List;
 public class Report{
     private List<Transaction> creditTransactions;
     private List<Transaction> debitTransactions;
-    private Integer totalCreditAmount;
-    private Integer totalDebitAmount;
-    private Integer netFlow;
+    private Double totalCreditAmount;
+    private Double totalDebitAmount;
+    private Double netFlow;
 
     public Report() {
-        this.totalCreditAmount = 0;
-        this.totalDebitAmount = 0;
-        this.netFlow = 0;
+        this.totalCreditAmount = 0.0;
+        this.totalDebitAmount = 0.0;
+        this.netFlow = 0.0;
     }
 
     public void calculateTotals() {
